Split debug_raw_processor.py into debug_*.py and run_*.py
